Crate qrcode

QRCode encoder

This crate provides a QR code and Micro QR code encoder for binary data.

use qrcode::QrCode;
use image::Luma;

// Encode some data into bits.
let code = QrCode::new(b"01234567").unwrap();

// Render the bits into an image.
let image = code.render::<Luma<u8>>().build();

// Save the image.
image.save("/tmp/qrcode.png").unwrap();

// You can also render it into a string.
let string = code.render()
    .light_color(' ')
    .dark_color('#')
    .build();
println!("{}", string);

Re-exports

pub use crate::types::Color;

pub use crate::types::EcLevel;

pub use crate::types::QrResult;

pub use crate::types::Version;

Modules

bits

The bits module encodes binary data into raw bits used in a QR code.

canvas

The canvas module puts raw bits into the QR code canvas.

ec

The ec module applies the Reed-Solomon error correction codes.

optimize

Find the optimal data mode sequence to encode a piece of data.

render

Render a QR code into image.

types

Structs

QrCode

The encoded QR code symbol.