Module qrcode::bits

The bits module encodes binary data into raw bits used in a QR code.

Structs

Bits
The Bits structure stores the encoded data for a QR code.

Enums

ExtendedMode
An "extended" mode indicator, includes all indicators supported by QR code beyond those bearing data.

Functions

encode_auto
Automatically determines the minimum version to store the data, and encode the result.