Enum qrcode::canvas::Module

pub enum Module {
    Empty,
    Light,
    Dark,
    LightUnmasked,
    DarkUnmasked,
}

The color of a module (pixel) in the QR code.

Variants

Empty

The module is empty.

Light

The module is light (white), and cannot be masked. This mainly refers to modules of functional patterns.

Dark

The module is dark (black), and cannot be masked. This mainly refers to modules of functional patterns.

LightUnmasked

The module is light (white), but not yet masked. This mainly refers to modules of data and error correction bits before masking.

DarkUnmasked

The module is dark (black), but not yet masked. This mainly refers to modules of data and error correction bits before masking.

Methods

impl Module

fn is_dark(&self) -> bool

Checks whether a module is dark.

fn mask(&self, should_invert: bool) -> Module

Apply a mask to the unmasked modules.

use qrcode::canvas::Module;

assert_eq!(Module::LightUnmasked.mask(true), Module::Dark);
assert_eq!(Module::DarkUnmasked.mask(true), Module::Light);
assert_eq!(Module::LightUnmasked.mask(false), Module::Light);
assert_eq!(Module::Dark.mask(true), Module::Dark);
assert_eq!(Module::Dark.mask(false), Module::Dark);