Module qrcode::optimize

Find the optimal data mode sequence to encode a piece of data.

Structs

Optimizer

Parser
QR code data parser to classify the input into distinct segments.

Segment
A segment of data committed to an encoding mode.

Functions

total_encoded_len
Computes the total encoded length of all segments.