Struct qrcode::optimize::Segment

pub struct Segment {
    pub mode: Mode,
    pub begin: usize,
    pub end: usize,
}

A segment of data committed to an encoding mode.

Fields

mode

The encoding mode of the segment of data.

begin

The start index of the segment.

end

The end index (exclusive) of the segment.

Methods

impl Segment

fn encoded_len(&self, version: Version) -> usize

Compute the number of bits (including the size of the mode indicator and length bits) when this segment is encoded.