Module qrcode::types

Enums

EcLevel
The error correction level. It allows the original information be recovered even if parts of the code is damaged.

Mode
The mode indicator, which specifies the character set of the encoded data.

QrError
QrError encodes the error encountered when generating a QR code.

Version
In QR code terminology, Version means the size of the generated image. Larger version means the size of code is larger, and therefore can carry more information.

Type Definitions

QrResult
QrResult is a convenient alias for a QR code generation result.