Enum qrcode::types::Version

pub enum Version {
    Normal(i16),
    Micro(i16),
}

In QR code terminology, Version means the size of the generated image. Larger version means the size of code is larger, and therefore can carry more information.

The smallest version is Version::Normal(1) of size 21×21, and the largest is Version::Normal(40) of size 177×177.

Variants

Normal

A normal QR code version. The parameter should be between 1 and 40.

Micro

A Micro QR code version. The parameter should be between 1 and 4.

Methods

impl Version

fn width(&self) -> i16

Get the number of "modules" on each size of the QR code, i.e. the width and height of the code.

fn fetch<T>(&self, ec_level: EcLevel, table: &[[T; 4]]) -> QrResult<T> where T: PartialEq + Default + Copy

Obtains an object from a hard-coded table.

The table must be a 44×4 array. The outer array represents the content for each version. The first 40 entry corresponds to QR code versions 1 to 40, and the last 4 corresponds to Micro QR code version 1 to 4. The inner array represents the content in each error correction level, in the order [L, M, Q, H].

If the entry compares equal to the default value of T, this method returns Err(QrError::InvalidVersion).

fn mode_bits_count(&self) -> usize

The number of bits needed to encode the mode indicator.

fn is_micro(&self) -> bool

Checks whether is version refers to a Micro QR code.